DEV: Refactor `DiscourseAutomation::DestroyAutomation` a bit
Small followup to 932bd6b.
